+++ src/routes/kmis/studentRoutes.js
@@ -7,6 +7,7 @@ const validate = require("../../middlewares/validate");
 const authMiddleware = require("../../middlewares/authMiddleware");
 const rateLimiter = require("../../middlewares/rateLimitMiddleware");
 const requireAbility = require("../../middlewares/requireAbility");
+const requirePermission = require("../../middlewares/requirePermission");
 
 router.use(
   rateLimiter,
@@ -26,6 +27,7 @@ router.get(
 
 router.post(
   "/sso/create",
+  requirePermission(["create.kmis_student"]),
   storeStudentValidator,
   validate,
   studentController.store
@@ -33,6 +35,7 @@ router.post(
 
 router.patch(
   "/sso/update/:id",
+  requirePermission(["update.kmis_student"]),
   updateStudentValidator,
   validate,
   studentController.update
@@ -40,21 +43,25 @@ router.patch(
 
 router.delete(
   "/sso/delete",
+  requirePermission(["delete.kmis_student"]),
   studentController.destroy
 );
 
 router.patch(
   "/sso/restore",
+  requirePermission(["restore.kmis_student"]),
   studentController.restore
 );
 
 router.patch(
   "/sso/deactivate",
+  requirePermission(["update.kmis_student"]),
   studentController.deactivateAccount
 );
 
 router.patch(
   "/sso/activate",
+  requirePermission(["update.kmis_student"]),
   studentController.activateAccount
 );
